discord-moderation-bot/requirements.txt
discord.py
detoxify
torch
asyncio
python-dotenv
